test: requests wait for schema cache load
* nix: add postgrest-test-big-schema command
