+++ .github/workflows/test.yaml
@@ -81,7 +81,7 @@ jobs:
         uses: ./.github/actions/setup-nix
         with:
           authToken: '${{ secrets.CACHIX_AUTH_TOKEN }}'
-          tools: tests.testSpec.bin tests.testIO.bin withTools.postgresql-${{ matrix.pgVersion }}.bin
+          tools: tests.testSpec.bin tests.testIO.bin tests.testBigSchema.bin withTools.postgresql-${{ matrix.pgVersion }}.bin
 
       - name: Run spec tests
         if: always()
@@ -90,6 +90,10 @@ jobs:
       - name: Run IO tests
         if: always()
         run: postgrest-with-postgresql-${{ matrix.pgVersion }} postgrest-test-io -vv
+
+      - name: Run IO tests on a big schema
+        if: always()
+        run: postgrest-with-postgresql-${{ matrix.pgVersion }} postgrest-test-big-schema -vv
 
 
   memory:
