+++ nix/tools/devTools.nix
@@ -76,6 +76,7 @@ let
         ${tests}/bin/postgrest-test-spec
         ${tests}/bin/postgrest-test-doctests
         ${tests}/bin/postgrest-test-io
+        ${tests}/bin/postgrest-test-big-schema
         ${style}/bin/postgrest-lint
         ${style}/bin/postgrest-style-check
       '';
